Add lib32-libpng, tbh not totally sure of its use
It was explicitly installed (apparently?) but can't be removed because
it's a dependency of another package.

According to `pactree -r lib32-libpng`

lib32-libpng
└─lib32-freetype2
  └─lib32-harfbuzz
    └─lib32-freetype2

Which gives a circular dependency such that when I try to remove
lib32-freetype2 it complains

```
$ sudo pacman -R lib32-freetype2
...
error: failed to prepare transaction (could not satisfy dependencies)
:: removing lib32-freetype2 breaks dependency 'lib32-freetype2' required by lib32-harfbuzz
:: removing lib32-freetype2 breaks dependency 'libfreetype.so=6-32' required by lib32-harfbuzz
```

What a weird circular dependency error, I've never seen something like
that before! Related, I did find this forum post on a similar dependency
loop of kio <--> kinit that helped me understand more of the matter.
https://bbs.archlinux.org/viewtopic.php?id=224884

I don't think it's so much of a big deal, apparently harfbuzz is a font
rendering library so keeping it in my system is probably not the worst
thing in the world.